+++ R/TernaryPlot.R
@@ -56,6 +56,16 @@
 #'  if axis labels are being clipped.
 #' @param col The colour for filling the plot; see \code{\link[graphics]{polygon}}.
 #' 
+#' @param panel.first An expression to be evaluated after the plot axes are
+#' set up but before any plotting takes place.
+#' This can be useful for drawing backgrounds, e.g. with [`ColourTernary()`]
+#' or [`HorizontalGrid()`].
+#' Note that this works by lazy evaluation: passing this argument from other
+#' plot methods may well not work since it may be evaluated too early.
+#' @param panel.last An expression to be evaluated after plotting has taken
+#' place but before the axes and box are added.  See the comments about
+#' `panel.first`.
+#' 
 #' @param grid.lines Integer specifying the number of grid lines to plot.
 #' @param grid.minor.lines Integer specifying the number of minor (unlabelled) 
 #'  grid lines to plot between each major pair.
@@ -120,7 +130,7 @@
 #' @importFrom graphics par plot polygon
 #' @export
 TernaryPlot <- function (atip = NULL, btip = NULL, ctip = NULL,
-                         alab = NULL, blab = NULL, clab = NULL, 
+                         alab = NULL, blab = NULL, clab = NULL,
                          lab.offset = 0.16, lab.col = NULL,
                          point = 'up', clockwise = TRUE,
                          xlim = NULL, ylim = NULL,
@@ -131,6 +141,7 @@ TernaryPlot <- function (atip = NULL, btip = NULL, ctip = NULL,
                          atip.pos = NULL, btip.pos = NULL, ctip.pos = NULL,
                          padding = 0.08,
                          col = NA,
+                         panel.first = NULL, panel.last = NULL,
                          grid.lines = 10, grid.col = 'darkgrey',
                          grid.lty = 'solid', grid.lwd = par('lwd'),
                          grid.minor.lines = 4, grid.minor.col = 'lightgrey',
@@ -145,6 +156,8 @@ TernaryPlot <- function (atip = NULL, btip = NULL, ctip = NULL,
                          ticks.lwd = axis.lwd, ticks.length = 0.025,
                          axis.col = 'black', ticks.col = grid.col,
                          ...) {
+  
+
   
   tern <- .TrianglePlot(
     atip = atip, btip = btip, ctip = ctip,
@@ -205,6 +218,8 @@ TernaryPlot <- function (atip = NULL, btip = NULL, ctip = NULL,
   options('.Last.triangle' = tern)
   
   .PlotBackground(tern)
+  
+  panel.first
 
   .PlotMinorGridLines(tern$grid.lines, tern$grid.minor.lines, 
                       col = tern$grid.minor.col,
@@ -219,6 +234,8 @@ TernaryPlot <- function (atip = NULL, btip = NULL, ctip = NULL,
   .PlotAxisTicks(tern)
   
   .PlotAxisLabels(tern)
+  
+  panel.last
   
   lapply(1:3, .AxisLines)
   lapply(1:3, .TitleAxis)
@@ -327,11 +344,11 @@ CoordinatesToXY <- function (coordinates) {
     } else if (is.numeric(coordinates)) {
       matrix(TernaryCoords(coordinates), nrow = 2)
     }
-  } else if (length(dims) == 2) {
-    which_dim <- if(dims[2] == 3) {
-      1
-    } else if (dims[1] == 3) {
-      2
+  } else if (length(dims) == 2L) {
+    which_dim <- if(dims[2] == 3L) {
+      1L
+    } else if (dims[1] == 3L) {
+      2L
     } else {
       stop("Coordinates must be ternary points")
     }
@@ -341,10 +358,23 @@ CoordinatesToXY <- function (coordinates) {
   }
 }
 
+#' @describeIn AddToTernary Add \link[graphics]{segments}
+#' @importFrom graphics segments
+#' @export
+TernarySegments <- function (fromCoordinates, toCoordinates = fromCoordinates,
+                             ...) {
+  fromXY <- CoordinatesToXY(fromCoordinates)
+  toXY <- CoordinatesToXY(toCoordinates)
+  
+  # Return:
+  segments(fromXY[1L, ], fromXY[2L, ], toXY[1L, ], toXY[2L, ], ...)
+}
+
 #' @describeIn AddToTernary Add  \link[graphics]{arrows}
 #' @importFrom graphics arrows
 #' @export
-TernaryArrows <- function (fromCoordinates, toCoordinates=fromCoordinates, ...) {
+TernaryArrows <- function (fromCoordinates, toCoordinates = fromCoordinates,
+                           ...) {
   fromXY <- CoordinatesToXY(fromCoordinates)
   toXY <- CoordinatesToXY(toCoordinates)
   
@@ -355,22 +385,30 @@ TernaryArrows <- function (fromCoordinates, toCoordinates=fromCoordinates, ...) 
 #' @describeIn AddToTernary Add \link[graphics]{lines}
 #' @importFrom graphics lines
 #' @export
-TernaryLines <- function (coordinates, ...) AddToTernary(lines, coordinates, ...)
+TernaryLines <- function (coordinates, ...) {
+  AddToTernary(lines, coordinates, ...)
+}
 
 #' @describeIn AddToTernary Add \link[graphics]{points}
 #' @importFrom graphics points
 #' @export
-TernaryPoints <- function (coordinates, ...) AddToTernary(points, coordinates, ...)
+TernaryPoints <- function (coordinates, ...) {
+  AddToTernary(points, coordinates, ...)
+}
 
 #' @describeIn AddToTernary Add \link[graphics:polygon]{polygons}
 #' @importFrom graphics polygon
 #' @export
-TernaryPolygon <- function (coordinates, ...) AddToTernary(polygon, coordinates, ...)
+TernaryPolygon <- function (coordinates, ...) {
+  AddToTernary(polygon, coordinates, ...)
+}
 
 #' @describeIn AddToTernary Add \link[graphics]{text}
 #' @importFrom graphics text
 #' @export
-TernaryText <- function (coordinates, ...) AddToTernary(text, coordinates, ...)
+TernaryText <- function (coordinates, ...) {
+  AddToTernary(text, coordinates, ...)
+}
 
 #' @describeIn AddToTernary Add points, joined by lines
 #' @importFrom graphics lines points